requirements.txt
factory_boy >= 2.12
scrapy >= 1.7
pytest >= 5.2
python-dotenv >= 0.20
tqdm >= 4.64.0
